refactor: simplify wheel API
