+++ crates/rattler_installs_packages/Cargo.toml
@@ -66,11 +66,11 @@ flate2 = "1.0.28"
 pyproject-toml = "0.9.0"
 async-once-cell = "0.5.3"
 configparser = "3.0.4"
-cacache = { version = "12.0.0", default-features = false, features = ["tokio-runtime", "mmap"] }
+cacache = { version = "13.0.0", default-features = false, features = ["tokio-runtime", "mmap"] }
 async-recursion = "1.0.5"
 fs-err = "2.11.0"
 fs_extra = "1.3.0"
-async_http_range_reader = "0.6.0"
+async_http_range_reader = "0.6.1"
 which = "6.0.0"
 
 [dev-dependencies]
