+++ crates/rattler_installs_packages/src/python_env/venv.rs
@@ -2,8 +2,8 @@
 //! Now just use the python venv command to create the virtual environment.
 //! Later on we can look into actually creating the environment by linking to the python library,
 //! and creating the necessary files. See: [VEnv](https://packaging.python.org/en/latest/specifications/virtual-environments/#declaring-installation-environments-as-python-virtual-environments)
-use crate::artifacts::wheel::{InstallPaths, UnpackWheelOptions, Wheel};
-use crate::artifacts::wheel::{UnpackError, UnpackedWheel};
+use crate::artifacts::wheel::Wheel;
+use crate::install::{install_wheel, InstallPaths, UnpackError, UnpackWheelOptions, UnpackedWheel};
 use crate::python_env::{
     system_python_executable, FindPythonError, ParsePythonInterpreterVersionError,
     PythonInterpreterVersion,
@@ -95,7 +95,8 @@ impl VEnv {
         wheel: &Wheel,
         options: &UnpackWheelOptions,
     ) -> Result<UnpackedWheel, UnpackError> {
-        wheel.unpack(
+        install_wheel(
+            wheel,
             &self.location,
             &self.install_paths,
             &self.python_executable(),
