+++ crates/rattler_installs_packages/src/wheel_builder/build_environment.rs
@@ -1,4 +1,4 @@
-use crate::artifacts::wheel::UnpackWheelOptions;
+use crate::install::UnpackWheelOptions;
 use crate::types::ArtifactFromSource;
 
 use crate::python_env::{PythonLocation, VEnv};
@@ -83,6 +83,7 @@ impl TempBuildEnvironment {
 
 // include static build_frontend.py string
 const BUILD_FRONTEND_PY: &str = include_str!("./wheel_builder_frontend.py");
+
 /// A build environment for building wheels
 /// This struct contains the virtualenv and everything that is needed
 /// to execute the PEP517 build backend hools
